resolvd Shir suggestions
