+++ packages/jira-adapter/src/filters/jsm_types_deploy_filter.ts
@@ -19,11 +19,14 @@ import { getChangeData, isInstanceChange, Change, InstanceElement } from '@salto
 import { defaultDeployChange, deployChanges } from '../deployment/standard_deployment'
 import { FilterCreator } from '../filter'
 import { JSM_DUCKTYPE_SUPPORTED_TYPES } from '../config/api_config'
+import { ASSESTS_SCHEMA_TYPE } from '../constants'
+import { getWorkspaceId } from '../workspace_id'
 
-const SUPPORTED_TYPES = new Set(Object.keys(JSM_DUCKTYPE_SUPPORTED_TYPES))
+const ASSETS_SUPPORTED_TYPES = [ASSESTS_SCHEMA_TYPE]
+const SUPPORTED_TYPES = new Set(Object.keys(JSM_DUCKTYPE_SUPPORTED_TYPES).concat(ASSETS_SUPPORTED_TYPES))
 
 const filterCreator: FilterCreator = ({ config, client }) => ({
-  name: 'jsmTypesFilter',
+  name: 'jsmDeployFilter',
   deploy: async (changes: Change<InstanceElement>[]) => {
     const { jsmApiDefinitions } = config
     if (!config.fetch.enableJSM || jsmApiDefinitions === undefined) {
@@ -32,17 +35,29 @@ const filterCreator: FilterCreator = ({ config, client }) => ({
         leftoverChanges: changes,
       }
     }
+
     const [jsmTypesChanges, leftoverChanges] = _.partition(
       changes,
       change =>
         SUPPORTED_TYPES.has(getChangeData(change).elemID.typeName)
         && isInstanceChange(change)
     )
+    const workspaceId = await getWorkspaceId(client)
+    const additionalUrlVars = workspaceId ? { workspaceId } : undefined
 
     const deployResult = await deployChanges(
       jsmTypesChanges,
       async change => {
-        await defaultDeployChange({ change, client, apiDefinitions: jsmApiDefinitions })
+        const typeDefinition = jsmApiDefinitions.types[getChangeData(change).elemID.typeName]
+        const deployRequest = typeDefinition.deployRequests ? typeDefinition.deployRequests[change.action] : undefined
+        const fieldsToIgnore = deployRequest?.fieldsToIgnore ?? []
+        await defaultDeployChange({
+          change,
+          client,
+          apiDefinitions: jsmApiDefinitions,
+          fieldsToIgnore,
+          additionalUrlVars,
+        })
       }
     )
     return { deployResult, leftoverChanges }
